Restructures example 3 into two cases.
One with hard coded mapping and one with ontology.
Also updated Readme
